# specs/openapi.json
{
	"openapi": "3.1.0",
	"info": {
		"title": "Stellar MVP API",
		"version": "0.1.0",
		"description": "API de Stellar MVP y servicios base."
	},
	"servers": [
		{
			"url": "http://127.0.0.1:3000",
			"description": "Development"
		}
	],
	"tags": [
		{
			"name": "health"
		},
		{
			"name": "stellar"
		},
		{
			"name": "items"
		},
		{
			"name": "agent"
		}
	],
	"paths": {
		"/": {
			"get": {
				"operationId": "redirectRootToDocs",
				"tags": ["health"],
				"summary": "Redirect to interactive API documentation",
				"responses": {
					"302": {
						"description": "Temporary redirect to /docs",
						"headers": {
							"location": {
								"schema": {
									"type": "string",
									"const": "/docs"
								}
							}
						}
					}
				}
			}
		},
		"/openapi.json": {
			"get": {
				"operationId": "getOpenApiSpec",
				"tags": ["health"],
				"summary": "Get canonical OpenAPI spec",
				"responses": {
					"200": {
						"description": "OpenAPI document",
						"content": {
							"application/json": {
								"schema": {
									"type": "object"
								}
							}
						}
					}
				}
			}
		},
		"/v1/health/live": {
			"get": {
				"operationId": "getLiveHealth",
				"tags": ["health"],
				"summary": "Liveness probe",
				"responses": {
					"200": {
						"description": "Process is alive",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/HealthResponse"
								},
								"examples": {
									"ok": {
										"value": {
											"status": "ok",
											"service": "stellarmvp"
										}
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/hello_api": {
			"get": {
				"operationId": "getStellarNetworkStatus",
				"tags": ["stellar"],
				"summary": "Check whether the Stellar network is configured",
				"responses": {
					"200": {
						"description": "Stellar network configuration status",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/StellarNetworkStatus"
								}
							}
						}
					}
				}
			}
		},
		"/v1/health/ready": {
			"get": {
				"operationId": "getReadiness",
				"tags": ["health"],
				"summary": "Readiness probe",
				"responses": {
					"200": {
						"description": "Readiness state",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ReadinessResponse"
								}
							}
						}
					}
				}
			}
		},
		"/v1/items": {
			"post": {
				"operationId": "createItem",
				"tags": ["items"],
				"summary": "Create an item",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateItemRequest"
							},
							"examples": {
								"demo": {
									"value": {
										"name": "demo",
										"metadata": {}
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Created item",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Item"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid service token",
						"headers": {
							"x-error-code": {
								"$ref": "#/components/headers/ErrorCodeHeader"
							}
						},
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"422": {
						"description": "Invalid request",
						"headers": {
							"x-error-code": {
								"$ref": "#/components/headers/ErrorCodeHeader"
							}
						},
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				},
				"security": [
					{
						"serviceToken": []
					}
				]
			}
		},
		"/v1/items/{itemId}": {
			"get": {
				"operationId": "getItem",
				"tags": ["items"],
				"summary": "Get an item",
				"parameters": [
					{
						"name": "itemId",
						"in": "path",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "Item",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Item"
								}
							}
						}
					},
					"401": {
						"description": "Missing or invalid service token",
						"headers": {
							"x-error-code": {
								"$ref": "#/components/headers/ErrorCodeHeader"
							}
						},
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"409": {
						"description": "Resource state conflict",
						"headers": {
							"x-error-code": {
								"$ref": "#/components/headers/ErrorCodeHeader"
							}
						},
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				},
				"security": [
					{
						"serviceToken": []
					}
				]
			}
		},
		"/v1/agent/search": {
			"post": {
				"operationId": "agentSearch",
				"tags": ["agent"],
				"summary": "Run a scoped agentic vector search",
				"description": "Executes a vector search on behalf of an authenticated agent. Requires the agent:search scope. Rate limited per agent token.",
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AgentSearchQuery"
							},
							"examples": {
								"demo": {
									"value": {
										"query": "wireless headphones",
										"limit": 10,
										"filters": {
											"merchant": "canonical"
										}
									}
								}
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Ranked search results",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/AgentSearchResponse"
								}
							}
						}
					},
					"401": {
						"description": "RFC 9457 problem response",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"403": {
						"description": "RFC 9457 problem response",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"422": {
						"description": "RFC 9457 problem response",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"429": {
						"description": "RFC 9457 problem response",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"503": {
						"description": "RFC 9457 problem response",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				},
				"security": [
					{
						"agentToken": []
					}
				]
			}
		},
		"/v1/agent/checkout": {
			"post": {
				"operationId": "agentCheckout",
				"tags": ["agent"],
				"summary": "Create a paid order via ACP x402 checkout",
				"description": "Creates an order on behalf of an authenticated agent. Requires the agent:checkout scope and a valid X-402-Payment-Token. When the token is missing or invalid the endpoint answers 402 Payment Required with an X-402-Challenge header for ACP x402 auto-negotiation. The order is persisted and marked paid only after the Stellar transaction is confirmed on-chain.",
				"parameters": [
					{
						"name": "X-402-Payment-Token",
						"in": "header",
						"required": false,
						"schema": {
							"type": "string"
						},
						"description": "ACP x402 payment token. Missing or invalid tokens receive a 402 challenge."
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AgentCheckoutRequest"
							},
							"examples": {
								"demo": {
									"value": {
										"itemId": "3fa85f64-5717-4562-b3fc-2c963f66afa6",
										"amount": "12.5000000",
										"currency": "USDC",
										"destination": "GAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAAWHF",
										"idempotencyKey": "agent-cart-1"
									}
								}
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Order created and paid on-chain",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/Order"
								}
							}
						}
					},
					"402": {
						"description": "Payment required or on-chain settlement failed. A 402 negotiation response carries the X-402-Challenge header.",
						"headers": {
							"X-402-Challenge": {
								"$ref": "#/components/headers/X402ChallengeHeader"
							},
							"x-error-code": {
								"$ref": "#/components/headers/ErrorCodeHeader"
							}
						},
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"403": {
						"description": "RFC 9457 problem response",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"422": {
						"description": "RFC 9457 problem response",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"429": {
						"description": "RFC 9457 problem response",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					},
					"503": {
						"description": "RFC 9457 problem response",
						"content": {
							"application/problem+json": {
								"schema": {
									"$ref": "#/components/schemas/Problem"
								}
							}
						}
					}
				},
				"security": [
					{
						"agentToken": []
					}
				]
			}
		},
		"/v1/payment-intents": {
			"post": {
				"operationId": "createPaymentIntent",
				"tags": ["payments"],
				"summary": "Create a Stellar Testnet payment intent from an approved quote",
				"parameters": [
					{
						"name": "x-principal-id",
						"in": "header",
						"required": true,
						"schema": { "type": "string", "minLength": 1, "maxLength": 128 }
					},
					{
						"name": "Idempotency-Key",
						"in": "header",
						"required": true,
						"schema": { "type": "string", "minLength": 8, "maxLength": 128 }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreatePaymentIntentRequest"
							}
						}
					}
				},
				"responses": {
					"201": {
						"description": "Intent created",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PaymentIntent" }
							}
						}
					},
					"200": {
						"description": "Idempotent replay",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PaymentIntent" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ProblemResponse" },
					"401": { "$ref": "#/components/responses/ProblemResponse" },
					"403": { "$ref": "#/components/responses/ProblemResponse" },
					"404": { "$ref": "#/components/responses/ProblemResponse" },
					"409": { "$ref": "#/components/responses/ProblemResponse" },
					"503": { "$ref": "#/components/responses/ProblemResponse" }
				},
				"security": [{ "serviceToken": [] }]
			}
		},
		"/v1/payment-intents/{intentId}": {
			"get": {
				"operationId": "getPaymentIntent",
				"tags": ["payments"],
				"summary": "Get and reconcile a payment intent",
				"parameters": [
					{
						"name": "intentId",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid" }
					},
					{
						"name": "x-principal-id",
						"in": "header",
						"required": true,
						"schema": { "type": "string", "minLength": 1, "maxLength": 128 }
					}
				],
				"responses": {
					"200": {
						"description": "Current intent state",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PaymentIntent" }
							}
						}
					},
					"401": { "$ref": "#/components/responses/ProblemResponse" },
					"403": { "$ref": "#/components/responses/ProblemResponse" },
					"404": { "$ref": "#/components/responses/ProblemResponse" },
					"409": { "$ref": "#/components/responses/ProblemResponse" },
					"503": { "$ref": "#/components/responses/ProblemResponse" }
				},
				"security": [{ "serviceToken": [] }]
			}
		},
		"/v1/payment-intents/{intentId}/submission": {
			"post": {
				"operationId": "submitPaymentIntentTransaction",
				"tags": ["payments"],
				"summary": "Verify and submit a wallet-signed Stellar transaction",
				"parameters": [
					{
						"name": "intentId",
						"in": "path",
						"required": true,
						"schema": { "type": "string", "format": "uuid" }
					},
					{
						"name": "x-principal-id",
						"in": "header",
						"required": true,
						"schema": { "type": "string", "minLength": 1, "maxLength": 128 }
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SubmitPaymentTransactionRequest"
							}
						}
					}
				},
				"responses": {
					"200": {
						"description": "Terminal result",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PaymentIntent" }
							}
						}
					},
					"202": {
						"description": "Submitted and pending reconciliation",
						"content": {
							"application/json": {
								"schema": { "$ref": "#/components/schemas/PaymentIntent" }
							}
						}
					},
					"400": { "$ref": "#/components/responses/ProblemResponse" },
					"401": { "$ref": "#/components/responses/ProblemResponse" },
					"403": { "$ref": "#/components/responses/ProblemResponse" },
					"404": { "$ref": "#/components/responses/ProblemResponse" },
					"409": { "$ref": "#/components/responses/ProblemResponse" },
					"503": { "$ref": "#/components/responses/ProblemResponse" }
				},
				"security": [{ "serviceToken": [] }]
			}
		}
	},
	"components": {
		"responses": {
			"ProblemResponse": {
				"description": "RFC 9457 problem response",
				"content": {
					"application/problem+json": {
						"schema": { "$ref": "#/components/schemas/Problem" }
					}
				}
			}
		},
		"headers": {
			"ErrorCodeHeader": {
				"description": "Stable error code. Matches problem.code.",
				"schema": {
					"type": "string",
					"pattern": "^SVC-[A-Z]{3,12}-[0-9]{4}$"
				}
			},
			"X402ChallengeHeader": {
				"description": "ACP x402 payment challenge negotiated by the agentic checkout endpoint.",
				"schema": {
					"type": "string"
				}
			}
		},
		"schemas": {
			"StellarNetworkStatus": {
				"type": "object",
				"required": ["stellarNetworkConfigured"],
				"properties": {
					"stellarNetworkConfigured": {
						"type": "boolean"
					}
				},
				"additionalProperties": false
			},
			"Problem": {
				"type": "object",
				"required": [
					"type",
					"title",
					"status",
					"code",
					"category",
					"detail_key",
					"behavior",
					"correlation",
					"occurred_at"
				],
				"properties": {
					"type": {
						"type": "string",
						"format": "uri"
					},
					"title": {
						"type": "string"
					},
					"status": {
						"type": "integer",
						"minimum": 400,
						"maximum": 599
					},
					"code": {
						"type": "string",
						"pattern": "^SVC-[A-Z]{3,12}-[0-9]{4}$"
					},
					"category": {
						"enum": [
							"VALIDATION",
							"AUTHN_AUTHZ",
							"DECISION_DENY",
							"STATE_CONFLICT",
							"DEPENDENCY",
							"INDETERMINATE",
							"INTERNAL"
						]
					},
					"detail_key": {
						"type": "string"
					},
					"behavior": {
						"$ref": "#/components/schemas/ErrorBehavior"
					},
					"correlation": {
						"$ref": "#/components/schemas/ErrorCorrelation"
					},
					"occurred_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"additionalProperties": false
			},
			"CreateItemRequest": {
				"type": "object",
				"required": ["name"],
				"properties": {
					"name": {
						"type": "string",
						"minLength": 1,
						"maxLength": 120
					},
					"metadata": {
						"type": "object",
						"additionalProperties": true,
						"default": {}
					}
				},
				"additionalProperties": false
			},
			"ErrorBehavior": {
				"type": "object",
				"required": [
					"retryable",
					"financial_effect",
					"human_action",
					"agent_hint",
					"retry_after_s"
				],
				"properties": {
					"retryable": {
						"enum": [
							"never",
							"safe_if_idempotent",
							"conditional",
							"after_reconcile"
						]
					},
					"financial_effect": {
						"enum": ["none", "unknown", "possible_partial"]
					},
					"human_action": {
						"enum": ["none", "approve", "step_up", "contact_support"]
					},
					"agent_hint": {
						"enum": [
							"NONE",
							"FIX_AND_RETRY",
							"RETRY_WITH_BACKOFF",
							"REQUEST_HUMAN_APPROVAL",
							"REQUEST_STEP_UP",
							"WAIT_FOR_SETTLEMENT",
							"ABORT_AND_REPORT"
						]
					},
					"retry_after_s": {
						"type": ["integer", "null"],
						"minimum": 0
					}
				},
				"additionalProperties": false
			},
			"ErrorCorrelation": {
				"type": "object",
				"required": ["trace_id"],
				"properties": {
					"trace_id": {
						"type": "string"
					}
				},
				"additionalProperties": false
			},
			"HealthResponse": {
				"type": "object",
				"required": ["status", "service"],
				"properties": {
					"status": {
						"const": "ok"
					},
					"service": {
						"type": "string"
					}
				},
				"additionalProperties": false
			},
			"Item": {
				"type": "object",
				"required": ["id", "name", "metadata", "created_at"],
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"name": {
						"type": "string"
					},
					"metadata": {
						"type": "object",
						"additionalProperties": true
					},
					"created_at": {
						"type": "string",
						"format": "date-time"
					}
				},
				"additionalProperties": false
			},
			"ReadinessResponse": {
				"type": "object",
				"required": ["status", "service", "details"],
				"properties": {
					"status": {
						"enum": ["ok", "degraded"]
					},
					"service": {
						"type": "string"
					},
					"details": {
						"type": "object",
						"additionalProperties": true
					}
				},
				"additionalProperties": false
			},
			"AgentSearchQuery": {
				"type": "object",
				"required": ["query"],
				"properties": {
					"query": {
						"type": "string",
						"minLength": 1,
						"maxLength": 500
					},
					"limit": {
						"type": "integer",
						"minimum": 1,
						"maximum": 50,
						"default": 10
					},
					"filters": {
						"type": "object",
						"additionalProperties": {
							"type": "string"
						}
					}
				},
				"additionalProperties": false
			},
			"AgentSearchResult": {
				"type": "object",
				"required": ["id", "score", "payload"],
				"properties": {
					"id": {
						"type": "string"
					},
					"score": {
						"type": "number"
					},
					"payload": {
						"type": "object",
						"additionalProperties": true
					}
				},
				"additionalProperties": false
			},
			"AgentSearchResponse": {
				"type": "object",
				"required": ["results", "count"],
				"properties": {
					"results": {
						"type": "array",
						"items": {
							"$ref": "#/components/schemas/AgentSearchResult"
						}
					},
					"count": {
						"type": "integer",
						"minimum": 0
					}
				},
				"additionalProperties": false
			},
			"AgentCheckoutRequest": {
				"type": "object",
				"required": ["itemId", "amount", "currency", "destination"],
				"properties": {
					"itemId": {
						"type": "string",
						"format": "uuid"
					},
					"amount": {
						"type": "string",
						"pattern": "^[0-9]+(\\.[0-9]+)?$"
					},
					"currency": {
						"type": "string",
						"minLength": 1,
						"maxLength": 12
					},
					"destination": {
						"type": "string",
						"minLength": 1,
						"maxLength": 128
					},
					"idempotencyKey": {
						"type": "string",
						"minLength": 1,
						"maxLength": 128
					}
				},
				"additionalProperties": false
			},
			"Order": {
				"type": "object",
				"required": [
					"id",
					"itemId",
					"amount",
					"currency",
					"status",
					"transactionHash",
					"createdAt"
				],
				"properties": {
					"id": {
						"type": "string",
						"format": "uuid"
					},
					"itemId": {
						"type": "string",
						"format": "uuid"
					},
					"amount": {
						"type": "string"
					},
					"currency": {
						"type": "string"
					},
					"status": {
						"enum": ["pending_payment", "paid", "failed", "fulfilled"]
					},
					"transactionHash": {
						"type": ["string", "null"]
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"additionalProperties": false
			}
		},
		"CreatePaymentIntentRequest": {
			"type": "object",
			"required": ["quoteId"],
			"properties": {
				"quoteId": { "type": "string", "minLength": 1, "maxLength": 128 }
			},
			"additionalProperties": false
		},
		"PaymentLeg": {
			"type": "object",
			"required": ["purpose", "payTo", "amountAtomic"],
			"properties": {
				"purpose": { "const": "merchant" },
				"payTo": { "type": "string", "pattern": "^G[A-Z2-7]{55}$" },
				"amountAtomic": { "type": "string", "pattern": "^[1-9][0-9]*$" }
			},
			"additionalProperties": false
		},
		"PaymentIntent": {
			"type": "object",
			"required": [
				"intentId",
				"quoteId",
				"orderId",
				"quoteHash",
				"status",
				"networkPassphrase",
				"payerAddress",
				"assetCode",
				"assetIssuer",
				"assetDecimals",
				"totalAmountAtomic",
				"paymentLeg",
				"unsignedXdr",
				"transactionHash",
				"ledger",
				"expiresAt",
				"createdAt",
				"updatedAt"
			],
			"properties": {
				"intentId": { "type": "string", "format": "uuid" },
				"quoteId": { "type": "string" },
				"orderId": { "type": "string" },
				"quoteHash": { "type": "string", "pattern": "^[a-f0-9]{64}$" },
				"status": {
					"enum": [
						"awaiting_signature",
						"submitting",
						"submitted",
						"confirmed",
						"failed",
						"expired"
					]
				},
				"networkPassphrase": { "type": "string" },
				"payerAddress": { "type": "string", "pattern": "^G[A-Z2-7]{55}$" },
				"assetCode": { "const": "USDC" },
				"assetIssuer": { "type": "string", "pattern": "^G[A-Z2-7]{55}$" },
				"assetDecimals": { "type": "integer", "const": 7 },
				"totalAmountAtomic": { "type": "string", "pattern": "^[1-9][0-9]*$" },
				"paymentLeg": { "$ref": "#/components/schemas/PaymentLeg" },
				"unsignedXdr": { "type": "string" },
				"transactionHash": {
					"type": ["string", "null"],
					"pattern": "^[a-f0-9]{64}$"
				},
				"ledger": { "type": ["integer", "null"], "minimum": 1 },
				"expiresAt": { "type": "string", "format": "date-time" },
				"createdAt": { "type": "string", "format": "date-time" },
				"updatedAt": { "type": "string", "format": "date-time" }
			},
			"additionalProperties": false
		},
		"SubmitPaymentTransactionRequest": {
			"type": "object",
			"required": ["signedXdr"],
			"properties": {
				"signedXdr": { "type": "string", "minLength": 1, "maxLength": 100000 }
			},
			"additionalProperties": false
		},
		"securitySchemes": {
			"serviceToken": {
				"type": "http",
				"scheme": "bearer",
				"description": "Optional service-to-service token. Required on protected routes when SERVICE_TOKEN is configured."
			},
			"agentToken": {
				"type": "http",
				"scheme": "bearer",
				"description": "Agent bearer token registered in Redis with scopes and a max daily spend. Required on /v1/agent/* routes."
			}
		}
	}
}